Create attachments.js
This file currently successfully inserts the content of a single attachment of an email sent to joe@zetalogo.com that is stored in S3(avoid loggin into joe's account from the main application because this will delete the email from S3 and store it into the emails table in the database, leaving no content in the S3 bucket to repeatedly test attachment handling with). The code still suffers from a timing issue in that the res.send call will execute before the insertion query is completed (solution most likely found but still lower priority than sending/receiving attachments).